Remove REDIS_HOST/REDIS_PORT configuration appraoch
This settles on REDIS_URL as the means to configure this gem, this was
chosen as it is one natively understood by the Redis gem [1] and thus
more common.

This choice has been removed as part of a GOV.UK effort to remove the
confusion that can be caused by having two different configuration
approaches used across the platform.

[1]: https://github.com/redis/redis-rb